There are some entities which is present in ER diagram and not into this project structure or code , that is add here in this commit but yet to connect with frontend

+++ src/main/java/com/hotel/hotelmanagement/Repository/HotelRepository.java
@@ -1,11 +1,9 @@
 package com.hotel.hotelmanagement.Repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import java.util.List;
 import com.hotel.hotelmanagement.Entity.Hotel;
+import java.util.List;
 
-public interface HotelRepository extends JpaRepository<Hotel, Integer> {
-
+public interface HotelRepository extends JpaRepository<Hotel, Long> {
     List<Hotel> findByRating(Integer rating);
-
 }
